move choice to the separate mod
