+++ rust/vote-tx-v2/src/public_tx/mod.rs
@@ -1,27 +1,27 @@
 //! A Catalyst public vote transaction v2 object, structured following this
 //! [spec](https://input-output-hk.github.io/catalyst-libs/architecture/08_concepts/catalyst_voting/v2/#public-vote)
 
+mod choice;
+mod decoding;
+
 use std::ops::{Deref, DerefMut};
+
+pub use choice::Choice;
 
 use crate::{
     gen_tx::{GeneralizedTx, Uuid},
     Cbor,
 };
 
-mod decoding;
-
 /// A public vote tx struct.
 pub struct PublicTx<VoteDataT>(GeneralizedTx<Choice, Proof, PropId, VoteDataT>)
 where VoteDataT: for<'a> Cbor<'a>;
-
-/// A public voting choice struct.
-pub struct Choice(pub u64);
 
 /// A public voting proof struct, CBOR `undefined`.
 pub struct Proof;
 
 /// A public voting proposal id struct.
-pub struct PropId(pub Uuid);
+pub struct PropId(Uuid);
 
 impl<VoteDataT> Deref for PublicTx<VoteDataT>
 where VoteDataT: for<'a> Cbor<'a>
